Check if who is updating is author or SuperAdmin

+++ backendSpring/src/main/java/org/java/app/photoalbum/controller/PhotoController.java
@@ -7,6 +7,10 @@ import org.java.app.photoalbum.db.pojo.Photo;
 import org.java.app.photoalbum.db.serv.CategoryService;
 import org.java.app.photoalbum.db.serv.PhotoService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
@@ -16,6 +20,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.server.ResponseStatusException;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
 import jakarta.validation.Valid;
@@ -30,7 +35,21 @@ public class PhotoController {
 	private CategoryService categoryService;
 	
 	private String savePhoto(Photo photo, BindingResult bindingResult, Model model, 
-			RedirectAttributes ra, boolean isNew) {
+			RedirectAttributes ra, boolean isNew, @AuthenticationPrincipal UserDetails user) {
+		
+		if (!isNew) {
+		
+		UserDetails userPhoto = photo.getUser();
+		boolean isSuperAdmin = user.getAuthorities().stream().anyMatch(auth -> "SUPER_ADMIN".equals(auth.getAuthority()));
+		if (userPhoto != user && !isSuperAdmin)
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+
+		}
+		
+		else {
+			photo.setUser((org.java.app.photoalbum.auth.pojo.User) user);
+		}
+		
 		if (bindingResult.hasErrors()) {
 			model.addAttribute("categories", categoryService.findAll());
 			return "photos/photo-create";
@@ -44,6 +63,8 @@ public class PhotoController {
 			return "redirect:/photos/" + photo.getId();
 
 		}}
+	
+	
 	
 	@GetMapping public String getIndex(Model model,
 			@RequestParam(required = false) String title) {
@@ -83,8 +104,8 @@ public class PhotoController {
 	
 	@PostMapping("/create")
 	public String storePhoto(@Valid @ModelAttribute Photo photo, BindingResult bindingResult, Model model,
-			RedirectAttributes ra) {
-		return savePhoto(photo, bindingResult, model, ra, true);
+			RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
+		return savePhoto(photo, bindingResult, model, ra, true, user);
 	}
 	
 	@GetMapping("/update/{id}")
@@ -103,17 +124,27 @@ public class PhotoController {
 	
 	@PostMapping("/update/{id}")
 	public String updatePhoto(@Valid @ModelAttribute Photo photo, BindingResult bindingResult, Model model,
-			RedirectAttributes ra) {
-		return savePhoto(photo, bindingResult, model, ra, false);		
+			RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
+		return savePhoto(photo, bindingResult, model, ra, false, user);		
 	}
 	
 	@PostMapping("/delete/{id}")
-	public String deletePizza(@PathVariable int id, RedirectAttributes ra) {
+	public String deletePizza(@PathVariable int id, RedirectAttributes ra, @AuthenticationPrincipal UserDetails user) {
 		Optional<Photo> optPhoto = photoService.findById(id);
 		if (optPhoto.isEmpty()) {
 			return "redirect:/photos";
 		}
+		
 		Photo photo = optPhoto.get();
+
+		org.java.app.photoalbum.auth.pojo.User userPhoto = photo.getUser();
+		boolean isSuperAdmin = user.getAuthorities().stream().anyMatch(auth -> "SUPER_ADMIN".equals(auth.getAuthority()));
+
+		if (userPhoto != user && !isSuperAdmin)
+		    throw new ResponseStatusException(HttpStatus.FORBIDDEN); 
+
+		
+		
 		ra.addFlashAttribute("updateMessage", "Foto con ID: " + photo.getId() + " (" + photo.getTitle() + ") cancellata");
 		photoService.delete(photo);
 		return "redirect:/photos";
